fixed user search and right edge layout
fixed user search by removing the code to pop matchedUsers. this wasn’t
working and was unnecessary because matchedUsers array is cleared for
each query

also added a calc function to keep the widget-container from
maintaining its original width when the nav is moved down vertically
next to it for tablet and up.